This commit adds some tooling and the basic structure planned for VIPER.
The plan is to separate tool from logical step - therefore, the tool is split into wrappers and stages.
Wrappers are meant to drive concrete external software tools like rosetta or gromacs, whereas stages are meant to perform a logical step in the execution of VIPER, like predicting the tertiary structure of the peptide.
Stages therefore use wrappers to achieve their goal, but wrappers are intended to be fairly agnostic of what specific logical steps are taken by VIPER.
Simultaneously, the high level execution of VIPER (see VIPER.py) is intended to be fairly agnostic of what specific tool is used to achieve the logical steps.
The idea is that these intermediary stages between the external software tools and the high level execution flow of VIPER provide additional flexibility and modularity.