Made fragment joiner aware of 3D structure/coordinates of fragments and fixed fragment linking.
In Detail:
There was an edge case in fragment linking, where residues would not get renumbered correctly because of the inserted linker residues, this is now fixed.
The linker can now also be truncated, if required, if the linker is longer than the inter-fragment gap.

Fixed an issue in FragmentJoiner, where the aggregate fragment strength score would get calculated incorrectly if lookahead was required to create the fragment.
Updated FragmentJoiner, so that it now keeps track of the alpha carbon coordinates of both the N- and C-terminus of each fragment. The new strategy now
constructs "inclusion" radii around each end of the termini and only considers adding other fragments that are within the inclusion radius when extending
the current candidate. The inclusion radius can be configured, it currently is a function of the linker length times a fixed value (4.0 Angstrom).
The tool iteratively extends the candidate by repeatedly choosing the fragment with the highest score among all fragments that are within the inclusion
radii of each terminus of the current candidate until the length limit is reached. The tool may now also join a subsequence of the selected fragment, if it
could otherwise not be added because of it being too long.
Also kept a previous version of making the linking smarter using 3D information, which penalized the scores of fragments based on the distance of each fragment
to the highest scoring fragment.